Add order_items.product_id index and BigInteger for amount columns
- Add index on order_items.product_id for product archival queries
- Upgrade orders.total_cents to BigInteger for consistency with budget columns

+++ backend/src/models/orm/order.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, Text, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -20,7 +20,7 @@ class Order(Base):
     status: Mapped[str] = mapped_column(
         String(50), nullable=False, default="pending"
     )
-    total_cents: Mapped[int] = mapped_column(Integer, nullable=False)
+    total_cents: Mapped[int] = mapped_column(BigInteger, nullable=False)
     delivery_note: Mapped[str | None] = mapped_column(Text, nullable=True)
     admin_note: Mapped[str | None] = mapped_column(Text, nullable=True)
     expected_delivery: Mapped[str | None] = mapped_column(String(255), nullable=True)
